pyproject.toml
[project]
name = "kindle-ocr"
version = "0.1.0"
description = "Kindleの本を自動的にスクリーンショット撮影し、OCRでテキスト化してMarkdownとPDFに変換するツール"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.2.1",
    "pyautogui>=0.9.54",
    "pymupdf>=1.26.0",
    "pyobjc-framework-Vision>=10.0",
]

[project.scripts]
kindle-pdf = "kindle_to_pdf.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kindle_to_pdf"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
]
